In update-helper ask before using a cache file
* If -y is not specified and a cache file is found, ask the user before
using it.
* "Fix" the -c option's behaviour. With this commit, if -c is applied,
the system will use the provided cache file if present and will do a
debootstrap if the file does not exist. The result of debootstrap will
be saved in the default cache file provided by the HELPER_CACHE_FILE
variable. This way, we do not need to provide an extra "disable cache
file option". One can run the program with -c "" to get this behaviour.