Add mechanism for monitoring snf-image events
Add a utility initially written by vkoukis for monitoring the image
copy. This utility outputs notifications to a named pipe. An external
program can get informed for the image deployment progress by reading
the notifications from this named pipe. The monitoring mechanism
will be extended in the future to have other snf-image subparts
output notification to this named pipe too.