disklabel.py: Rename Disklabel cls to Disk
The Disk class hosts the partitioning info of a Disk, namely an MBR
instance as well as a Disklabel instance.

For disklabels define 2 different classes:

 * BSDDisklabel
 * OpenBSDDisklabel

The first represents the traditional *BSD disklabel and the latter the
altered version found in newer releases of OpenBSD.

For now, only OpenBSDDisklabel is implemented.