Don't let sgdisk align when creating partitions
In order to extend a partition we need to remove and recreate it. If the
partition was misaligned then the image is to blame. snf-image is not
allowed to change the starting sector of the partition since this will
destroy the file system it hosts. In order to make sure this does never
happen we need to set the sector alignment multiple to 1.