Move DisableRemoteDesktopConnections to prio 41
Move DisableRemoteDesktopConnections from priority 40
to priority 41, ensuring it runs after 40InstalUnattend.in.

This makes the dependency between the installed answer file and
the DisableRemoteDesktopConnections task explicit:
The task assumes that RDP connections will be re-enabled via
an appropriate <RunSynchronousCommand> entry in the answer file,
which must already exist.

Making the dependency explicit, allows making the process more robust
in the future: The DisableRemoteDesktopConnections task should not
blindly assume that a potentially Image-specific unattend.xml file
contains the specific <RunSynchronousCommand> entry it requires,
but may insert it explicitly, since the answer file is bound to have
been detected or installed via the the previously executed
40InstallUnattend task.