kvm: Support various target device types
Old QEMU versions used -drive file=<file>,iface=<iface> in order
to define a virtual disk. Until now snf-image was using this
deprecated method and launched the helper using the virtio
interface. Here we use the new interface which goes down to:

 -drive file=<file>,iface=none,id=disk0
 -device <driver>,drive=disk0

Ganeti exports the disk type to be used as INSTANCE_HV_disk_type.
Some of the allowed values are: paravirtual, scsi, scsi-hd,
scsi-generic, scsi-block, etc.

In case of paravirtual device type the <driver> will be
"virtio-blk-pci" and the helper will see the disk as /dev/vdb.

If the disk type is scsi'ish then the helper will see the device as
/dev/sda and the <driver> will be:

  - scsi-hd for scsi
  - the same as disk type otherwise

Note that in case of a scsi'ish disk type an extra SCSI controller
(virtio-scsi-pci) is added to the QEMU command line.

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>