Add check_yes_no() to test boolean img properties
Use check_yes_no() to test if a boolean property is set.

Although the documentation states that to set a boolean image property
you need to assign the "yes" value to it, this function will accept
"yes", "true", 1, "on" and "set" in a case-insensitive way and reject
"no", "false", 0, "off" and "unset".

An empty or not-set variable is treated as false.

An unknown value will raise a warning but will be accepted. This is done
to protect the users because prior to this commit, in some cases we only
tested if an image property had a non-empty value.

This resolves #80