Add new os_answer_file image property
This can be used to specify a URI hosting a Windows answer file to be
injected to the image during deployment. Depending on the image (windows
or windows-legacy) this file can either be an xml file (unattend.xml) or
an inf one (sysprep.inf). The file gets higher priority than the file
that may be hosted inside the image.