Support userspace-only disks
Commits e147e00 and c4f1bff in Ganeti add support for userspace-only
disks. That means that we could have an instance without a local
block device (e.g. using the embedded support in QEMU for rbd by
passing -drive file=rbd:pool/image_name without the need of a
/dev/rdbX device).

This patch does the following:

Introduce a find_disk() helper method that first checks the
DISK_IDX_PATH and if it is not exported then it returns
DISK_IDX_URI.

Introduce a losetup_disk() helper method that is used during
export/import and in case we have an ntfsdump/extdump image type.

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>