Make the process of image copying modular
With this patch, snf-image supports a modular interface for copying
data from the provided image to the VM's actual disk.

For fetching image data, a variety of different source backends are
supported: `file', `network', `pithos', and `null'. For copying
these data to the VM's disk, this patch introduces the `file'
destination backend, i.e. for local block devices or files. Each
backend corresponds to a separate script.

In a nutshell, the copy procedure boils down to:

1) Parse IMG_ID and get the source backend type.
2) Parse DISK_0_PATH or DISK_0_URI and get the destination backend type.
3) Check if the corresponding scripts exist.
4) Invoke the source script with the `-s' option to get the image size.
5) Create a process chain of the source script, the monitor script,
   and the destination script "connected" with pipes.

The above steps will eventually fetch data from the image, pass them
to the monitor, and finally dump them to the VM's disk.

This patch also moves handling of file system dumps to the
script of the `file` destination backend.

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>