Replace name in md5sum file with the image name
snf-image-update-helper will check the MD5 sum of the downloaded helper
image file against the one found inside the downloaded .md5sum file.
Before doing this, replace the name found inside the .md5sum file with
the one of the downloaded image. They may be different due to the fact
that curl will construct the file name from the URL when downloading.
Till now, we used to put the name of the cdn.dev.grnet.gr URL inside the
md5sum file instead of the actual one.