Add a new auth_keys OS parameter
This parameter can be used to inject authorization keys to the VM for
the same users for whom a password change will be performed.

Prior to this commit, the same functionality could be achieved by
using the img_personality parameter to append the needed authorized_keys
files to the instance, but this is not the correct way to do it. Only
snf-image which does the deployment may know the actual path to the
authorized keys files.