Fix a bug where qemu-nbd cleanup was not performed
The problem was that the qemu-nbd cleanup was registered in map_uri()
were the qemu_nbd connection was performed. The function was called by
find_disk() which was called by the main shell like this:
disk=$(find_disk 0)

This is perfectly legal, but the functions runs in a sub-shell which
means that add_cleanup() will add the cleanup task to the CLEANUP array
of the sub-shell, which is never called.

This resolves #99