Add new inst_properties OS parameter
This can be used to overwrite keys defined in the img_properties OS
parameter. There are cases were the image properties do not reflect
the instance's state and need to be update. For example, a user may
provide cloud-init use-data that create or remove users. snf-image
cannot determine that the list of users, whose password needs to be
set, has changed, unless the user updates the USERS list.