Turn the back-ends to isolated modules
Write independent executable modules under backends/src that work with a
common API for the 4 different img_id types (local, network, pithos,
null) supported by snf-image. They expect to receive a URL as argument
and the optional arguments -s or -p. If -p is used, the module will
return whether it knows how to handle the provided URL. If -s is used
the size of the image is returned. This way, snf-image itself becomes
agnostic to the supported image back-ends. Every time it needs to fetch
an image, it can iterate over the list of available modules to find a
module that can handle the image ID. This design is extendable. The
users may provide new back-end modules to handle different types of
image back-ends.

Move image back-end options out of /etc/default/snf-image and put them
to module specific configuration files under
/etc/snf-image/backends/src. All pithos options should go to the
configuration of the pithos module. The IMAGE_DIR option should be in
the local module configuration.

Move /etc/default/snf-image to /etc/snf-image/snf-image.conf