Add priority and disable mechanism to back-ends
For every back-end the file <name>.priority and <name>.disabled can be
defined in the back-end configuration directory.

The priority file should contain a 2-digits number that defines a
priority. Modules with higher priority will be examined first when
trying to determine which back-end is suitable to handle a URL. If a
priority file is not present or has an invalid content, the modules
priority will be 50.

If the disable file is present, snf-image will not examine this module
at all.