Create destination back-ends
Use the same concept as the source back-ends to create destination
back-ends. The dst back-ends are modules that expect to receive data in
their stdin to transfer them to the ganeti target disk. snf-image will
probe the available destination back-ends and will pick the one that
supports the provided disk URI. The priorities and disable mechanism
work the same way as with the source back-ends.