Use splice() to transfer data btw back-ends
Till now, the copy-monitor.py helper script expected an program
(e.g. dd) as argument, it would execute it as child process and monitor
the bytes read by it. This worked well because the whole image copying
was performed in a single step.

Now that we pipe the output of the source back-end to the input of the
destination back-end, doing the image copy like this:

src | copy-monitor.py dd | dst

is inefficient. The pipe buffer is by default 64K and no matter how you
set the size of dd's input and output buffers, the copy is performed in
reduced speed.

To overcome this, instead of using an external program, make
copy-monitor.py move the input data to output using splice() after
growing the pipe capacity.